refactor:追加したコードをscore直下に移動しました．冗長なコードをカプセル化しました
