+++ typing-server/api/handler/score.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 	"github.com/su-its/typing/typing-server/api/service"
 	"github.com/su-its/typing/typing-server/domain/model"
@@ -78,4 +79,44 @@ func PostScore(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.WriteHeader(http.StatusCreated)
+}
+
+func GetMyScoreRanking(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+
+	userID := chi.URLParam(r, "user-id")
+	if userID == "" {
+		http.Error(w, "user-id is required", http.StatusBadRequest)
+		return
+	}
+
+	userUUID, err := uuid.Parse(userID)
+	if err != nil {
+		http.Error(w, "Invalid user-id", http.StatusBadRequest)
+		return
+	}
+
+	sortBy := r.URL.Query().Get("sort_by")
+	if sortBy == "" {
+		sortBy = "keystrokes"
+	}
+
+	currentRank, err := service.GetMyScoreRanking(ctx, entClient, userUUID, sortBy)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	response := struct {
+		CurrentRank int `json:"current-rank"`
+	}{
+		CurrentRank: currentRank,
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	err = json.NewEncoder(w).Encode(response)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
 }
